src/main/java/springAngularApp/system/configuration/SerializableResourceBundleMessageSource.java
package springAngularApp.system.configuration;

import org.springframework.context.support.ReloadableResourceBundleMessageSource;
import org.springframework.stereotype.Component;

import java.util.Locale;
import java.util.Properties;

@Component
public class SerializableResourceBundleMessageSource extends ReloadableResourceBundleMessageSource {

    public Properties getAllProperties(Locale locale) {
        clearCacheIncludingAncestors();
        setBasenames("classpath:/messages", "classpath:/users/messages"); //TODO Each module should have its own file to define resource
        PropertiesHolder propertiesHolder = getMergedProperties(locale);
        return propertiesHolder.getProperties();
    }

}